xe: jit: move dsl::kernel_t to IR
