+++ src/gpu/intel/jit/dsl/decl.hpp
@@ -46,6 +46,8 @@ static type_t bf16 = type_t::bf16();
 using expr_t = jit::expr_t;
 using type_t = jit::type_t;
 
+using kernel_t = jit::kernel_t;
+
 } // namespace dsl
 } // namespace jit
 } // namespace intel
